fix form request

+++ HRM/src/java/model/LeaveFormRequest.java
@@ -10,7 +10,7 @@ public class LeaveFormRequest extends FormRequest {
 
     private Date startDate;
     private Date endDate;
-    private Double totalDays;
+    private Integer totalDays;
 
     public Date getStartDate() { return startDate; }
     public void setStartDate(Date startDate) { this.startDate = startDate; }
@@ -18,6 +18,6 @@ public class LeaveFormRequest extends FormRequest {
     public Date getEndDate() { return endDate; }
     public void setEndDate(Date endDate) { this.endDate = endDate; }
 
-    public Double getTotalDays() { return totalDays; }
-    public void setTotalDays(Double totalDays) { this.totalDays = totalDays; }
+    public Integer getTotalDays() { return totalDays; }
+    public void setTotalDays(Integer totalDays) { this.totalDays = totalDays; }
 }
